admin_tools/pyproject.toml
[project]
name = "notes2notion-admin"
version = "1.0.0"
description = "Admin tools for Notes2Notion license key management"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "tabulate>=0.9.0",
]

[project.scripts]
license-manager = "license_manager:main"
